refactor: eliminate routing quality regressions
